+++ server/src/main/java/org/opensearch/search/backpressure/settings/SearchBackpressureSettings.java
@@ -63,6 +63,11 @@ public class SearchBackpressureSettings {
         Defaults.CANCELLATION_RATIO,
         0.0,
         1.0,
+        value -> {
+            if (value <= 0.0) {
+                throw new IllegalArgumentException("search_backpressure.cancellation_ratio must be greater than zero");
+            }
+        },
         Setting.Property.Deprecated,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
@@ -79,6 +84,12 @@ public class SearchBackpressureSettings {
         "search_backpressure.cancellation_rate",
         Defaults.CANCELLATION_RATE,
         0.0,
+        Double.MAX_VALUE,
+        value -> {
+            if (value <= 0.0) {
+                throw new IllegalArgumentException("search_backpressure.cancellation_rate must be greater than zero");
+            }
+        },
         Setting.Property.Deprecated,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope