+++ server/src/main/java/org/opensearch/search/backpressure/settings/SearchShardTaskSettings.java
@@ -50,6 +50,11 @@ public class SearchShardTaskSettings {
         SearchBackpressureSettings.SETTING_CANCELLATION_RATIO,
         0.0,
         1.0,
+        value -> {
+            if (value <= 0.0) {
+                throw new IllegalArgumentException("search_backpressure.search_shard_task.cancellation_ratio must be greater than zero");
+            }
+        },
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
     );
@@ -63,6 +68,12 @@ public class SearchShardTaskSettings {
         "search_backpressure.search_shard_task.cancellation_rate",
         SearchBackpressureSettings.SETTING_CANCELLATION_RATE,
         0.0,
+        Double.MAX_VALUE,
+        value -> {
+            if (value <= 0.0) {
+                throw new IllegalArgumentException("search_backpressure.search_shard_task.cancellation_rate must be greater than zero");
+            }
+        },
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
     );